+++ src/Scenes/Splash.cpp
@@ -5,7 +5,8 @@
 #include <string>
 #include "../../includes/Scenes/Splash.h"
 #include "../../includes/Scenes/SceneManager.h"
-#include "../../includes/Scenes/Game.h"
+#include "../../includes/Scenes/Mainmenu.h"
+#include "../../includes/Constants.h"
 
 SplashScreen::SplashScreen(SceneManager* sm) : Scene(sm) {}
 
@@ -16,8 +17,8 @@ void SplashScreen::draw() {
     DrawText("SZ", (GetScreenWidth() - titleWidth) / 2, GetScreenHeight() / 2, 90, YELLOW);
 
     // Draw Subtitle "A Game by Team Zero" Centered
-    int subtitleWidth = MeasureText("v 1.5.5", 20);
-    DrawText("v 1.5.5", (GetScreenWidth() - subtitleWidth) / 2, GetScreenHeight() / 2 + 90, 20, WHITE);
+    int subtitleWidth = MeasureText(VERSION_STRING, 20);
+    DrawText(VERSION_STRING, (GetScreenWidth() - subtitleWidth) / 2, GetScreenHeight() / 2 + 90, 20, WHITE);
 }
 
 void SplashScreen::update([[maybe_unused]] const float dt) {
@@ -25,7 +26,7 @@ void SplashScreen::update([[maybe_unused]] const float dt) {
 
     // Transition to next scene after 2.5 seconds
     if (time > 2.5f) {
-        sceneManager->changeScene(std::make_unique<Game>(sceneManager));
+        sceneManager->changeScene(std::make_unique<MainMenu>(sceneManager));
     }
 }